pfftools: extract seen_message_ids into a dedicated module
The deduplication hash table was embedded as static functions in
export_handle.c alongside unrelated MIME logic.  Moving it to its own
translation unit separates concerns and makes the interface available
without exposing internals.

- move seen_message_ids_table type definition, hash, free, add,
  contains, load, and save from export_handle.c to seen_message_ids.c,
  promoting static functions to extern so the header can declare them
- add seen_message_ids.h with the opaque typedef and function prototypes
- include seen_message_ids.h in export_handle.h; drop the now-redundant
  forward typedef from export_handle.h
- add seen_message_ids.c and seen_message_ids.h to pffexport_SOURCES in
  Makefile.am
- replace hand-rolled base64 loop in mime_base64_write_attachment with
  libuna_base64_triplet_copy_from_byte_stream and
  libuna_base64_triplet_copy_to_base64_stream, removing the local
  mime_b64_chars table and its manual bit-twiddling
